refactor: domain 내 클래스들 패키지 리팩터링
- domain 은 Entity만 들어가도록 한다
- dto.* 는 pojo 객체가 들어가도록 한다
