+++ src/main/java/com/api/cossc/dto/oauth/KakaoOAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package com.api.cossc.domain;
+package com.api.cossc.dto.oauth;
 
 import java.util.HashMap;
 import java.util.Map;
