+++ src/main/java/com/api/cossc/dto/oauth/OAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package com.api.cossc.domain;
+package com.api.cossc.dto.oauth;
 
 import java.util.Map;
 import lombok.Getter;
