+++ src/main/java/com/api/cossc/dto/oauth/UserRole.java
@@ -1,4 +1,4 @@
-package com.api.cossc.domain;
+package com.api.cossc.dto.oauth;
 
 
 public enum UserRole {
